+++ record-presentation/src/main/java/com/teamhy2/record/RecordScreen.kt
@@ -5,18 +5,13 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -28,12 +23,9 @@ import com.hongikyeolgong2.calendar.model.StudyDay
 import com.hongikyeolgong2.calendar.presentation.Hy2Calendar
 import com.teamhy2.designsystem.common.HY2CircularLoading
 import com.teamhy2.designsystem.ui.theme.BackgroundBlack
-import com.teamhy2.designsystem.ui.theme.Gray100
-import com.teamhy2.designsystem.ui.theme.Gray300
-import com.teamhy2.designsystem.ui.theme.HY2Theme
-import com.teamhy2.designsystem.ui.theme.HY2Typography
 import com.teamhy2.designsystem.util.compositionlocal.LocalShowSnackBar
 import com.teamhy2.designsystem.util.compositionlocal.LocalTracker
+import com.teamhy2.record.components.DatePanel
 import com.teamhy2.record.components.StudyDurationCard
 import com.teamhy2.record.components.StudyDurationCardType
 import com.teamhy2.record.domain.model.StudyDuration
@@ -141,52 +133,6 @@ fun RecordScreen(
             onNextMonthClick = onNextMonthClick,
             onDayClicked = onDayClicked,
             selectedDay = recordState.selectedStudyDay?.studyDay,
-        )
-    }
-}
-
-@Composable
-private fun DatePanel(
-    date: String,
-    hour: Int,
-    minute: Int,
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = modifier.fillMaxWidth(),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        Text(
-            text = date,
-            style = HY2Typography().body04,
-            color = Gray300,
-        )
-        Spacer(Modifier.height(6.dp))
-        Text(
-            text = "${hour}H ${minute}M",
-            style =
-                HY2Typography().body01.copy(
-                    brush =
-                        Brush.horizontalGradient(
-                            listOf(
-                                Color(0xFFD3D6E0),
-                                Color(0x99D3D6E0),
-                            ),
-                        ),
-                ),
-            color = Gray100,
-        )
-    }
-}
-
-@Preview(showBackground = true, backgroundColor = 0xFF000000)
-@Composable
-private fun DatePanelPreview() {
-    HY2Theme {
-        DatePanel(
-            date = "February 22, 2025",
-            hour = 3,
-            minute = 24,
         )
     }
 }
